+++ backend/src/game/types/key-data.interface.ts
